% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{compute_state_br}
\alias{compute_state_br}
\title{Title}
\usage{
compute_state_br(data, scales, keep_state = NULL)
}
\arguments{
\item{keep_state}{}
}
\description{
Title
}
\examples{
library(dplyr)
#brasil_flat |> rename(fips = FIPS) |> compute_state_br() |> head()
#brasil_flat |> rename(fips = FIPS) |> compute_state_br(keep_state = "Ashe")
}
