% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{compute_state_brasil_stamp}
\alias{compute_state_brasil_stamp}
\title{Title}
\usage{
compute_state_brasil_stamp(data, scales, keep_state = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#nc_flat |> rename(fips = FIPS) |> compute_state_brasil() |> head() |> str()
#nc_flat |> rename(fips = FIPS) |> compute_state_brasil(keep_state = "Ashe")
}
