% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{geom_label_brasil_state}
\alias{geom_label_brasil_state}
\title{Title}
\usage{
geom_label_brasil_state(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

library(ggplot2)
brasil_flat \%>\%
ggplot() +
aes(state = state) +
geom_sf_statebrasil() +
aes(fill = region, label = state) +

 geom_label_brasil_state(lineheight = .7,
 size = 2, check_overlap= TRUE,
 color = "oldlace")
}
